fix bug

+++ app/src/main/java/com/Meditation/Sounds/frequencies/lemeor/Extentions.kt
@@ -41,7 +41,7 @@ val max = MutableLiveData<Long>(0)
 val duration = MutableLiveData<Long>()
 
 //for track add to program
-var trackIdForProgram: Int? = -29000
+var trackIdForProgram: Double? = -29000.0
 var albumIdBackProgram: Int? = -1
 var categoryIdBackProgram: Int? = -1
 var typeBack: String = Constants.TYPE_ALBUM
